pyproject.toml
[project]
name = "greek-semantic-ui"
version = "0.1.0"
description = "Interactive UI for exploring Greek semantic embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.10.0",
    "transformers[torch]>=5.1.0",
    "sentence-transformers>=5.2.2",
    "pygame>=2.6.1",
    "numpy>=1.24.0",
    "ipykernel>=7.2.0",
    "pandas<3",
    "pyarrow<21",
    "streamlit>=1.30.0",
    "altair>=6.0.0",
    "scikit-learn>=1.4",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "faiss-cpu>=1.13.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
